Squashed 'kbuild/' changes from 89f3070..aac4c82
aac4c82 kmk_cc_exec.c: Make sure the variable doesn't get modified while we're running the expansion program. It'll screw up our string references.
4446038 [update] import upstream svn r2683
965762c Add .gitignore for binaries

git-subtree-dir: kbuild
git-subtree-split: aac4c82a1eb398bbf5a6ba712f04af2f8b03da8c

+++ src/kmk/kmkbuiltin.c
@@ -1,4 +1,4 @@
-/* $Id: kmkbuiltin.c 2413 2010-09-11 17:43:04Z bird $ */
+/* $Id$ */
 /** @file
  * kMk Builtin command execution.
  */
@@ -119,7 +119,7 @@ int kmk_builtin_command(const char *pszCmd, char ***ppapszArgvToSpawn, pid_t *pP
         argv[argc] = malloc(cch + 1);
         if (!argv[argc])
         {
-            printf("kmk_builtin: out of memory. argc=%d len=%d\n", argc, pszEnd - pszCmd + 1);
+            printf("kmk_builtin: out of memory. argc=%d len=%d\n", argc, (int)(pszEnd - pszCmd + 1));
             break;
         }
         memcpy(argv[argc], pszCmd, cch);
